Separated token and login endpoints
